+++ packages/ko/src/actions/build.ts
@@ -2,6 +2,7 @@ import webpack, { Configuration } from 'webpack';
 import Service from '../core/service';
 import WebpackConfig from '../webpack';
 import { ActionFactory } from './factory';
+import TerserWebpackPlugin from 'terser-webpack-plugin';
 
 import { ESBuildMinifyPlugin } from 'esbuild-loader';
 import { ICliOptions } from '../types';
@@ -17,14 +18,23 @@ class Build extends ActionFactory {
     const extraConfig = {
       optimization: {
         minimize: true,
-        minimizer: this.service.config.experiment?.minimizer
-          ? [
-              new ESBuildMinifyPlugin({
-                target: 'es2015',
-                css: true,
-              }),
-            ]
-          : ['...'],
+        minimizer: [
+          new TerserWebpackPlugin({
+            terserOptions: {
+              compress: {
+                ...this.service.config.experiment?.compress,
+              },
+            },
+          }),
+          ...(this.service.config.experiment?.minimizer
+            ? [
+                new ESBuildMinifyPlugin({
+                  target: 'es2015',
+                  css: true,
+                }),
+              ]
+            : ['...']),
+        ],
       },
     } as Configuration;
     const ret = this.webpackConfig.merge(extraConfig);
@@ -33,6 +43,10 @@ class Build extends ActionFactory {
       context: ret.plugins,
     });
     ret.plugins = plugins;
+    await this.service.apply({
+      key: this.service.hookKeySet.MODIFY_WEBPACK,
+      context: ret,
+    });
     return ret;
   }