+++ packages/ko/src/actions/dev.ts
@@ -6,6 +6,8 @@ import Service from '../core/service';
 import WebpackConfig from '../webpack';
 import ActionFactory from './factory';
 import { ICliOptions } from '../types';
+import detect from 'detect-port';
+import inquirer from 'inquirer';
 
 class Dev extends ActionFactory {
   private webpackConfig: WebpackConfig;
@@ -15,7 +17,7 @@ class Dev extends ActionFactory {
 
   private get devServerConfig(): DevServerConfiguration {
     const { serve, publicPath } = this.service.config;
-    const { host, port, proxy, staticPath } = serve;
+    const { host, port, proxy, staticPath, historyApiFallback = false } = serve;
     return {
       port,
       host,
@@ -30,8 +32,8 @@ class Dev extends ActionFactory {
       allowedHosts: 'all',
       client: {
         overlay: false,
-        logging: 'none',
       },
+      historyApiFallback,
     };
   }
 
@@ -57,10 +59,18 @@ class Dev extends ActionFactory {
       lazyCompilation:
         speedUp && disableLazyImports
           ? {
+              entries: false,
               imports: false,
             }
-          : speedUp,
+          : {
+              entries: false,
+              imports: true,
+            },
     };
+    await this.service.apply({
+      key: this.service.hookKeySet.MODIFY_WEBPACK,
+      context: ret,
+    });
     return ret;
   }
 
@@ -83,13 +93,44 @@ class Dev extends ActionFactory {
     this.service.commander.bindAction(cmdName, this.action.bind(this));
   }
 
+  private async changePort(newPort: number, port: number) {
+    const question = {
+      type: 'confirm',
+      name: 'changePort',
+      message: `port: ${port} has been used，use new port ${newPort} instead?`,
+      default: true,
+    };
+    const answer = await inquirer.prompt([question]);
+    if (answer.changePort) {
+      return newPort;
+    }
+    this.errorStdout(`so sorry, ${port} already in use！！`);
+    process.exit(0);
+  }
+
+  private async checkPort(port: number) {
+    const newPort = await detect(port);
+    if (newPort === port) {
+      return newPort;
+    }
+    const isInteractive = process.stdout.isTTY;
+    if (isInteractive) {
+      return this.changePort(newPort, port);
+    }
+  }
+
   protected async action(cliOpts: ICliOptions) {
     process.title = 'ko-dev';
     process.env.NODE_ENV = 'development';
     this.service.freezeCliOptsWith(cliOpts);
     const config = await this.generateConfig();
+    const port = config.devServer?.port as number;
+    const newPort = (await this.checkPort(port)) as number;
     const compiler = Webpack(config);
-    const devServer = new WebpackDevServer(config.devServer, compiler);
+    const devServer = new WebpackDevServer(
+      { ...config.devServer, port: newPort },
+      compiler
+    );
     await devServer.start();
     const exitProcess = (callback?: () => void) => () => {
       callback && callback();