+++ packages/ko/src/webpack/loaders/style.ts
@@ -7,6 +7,7 @@ import { getResolvePath } from '../../utils';
 import { IWebpackOptions } from '../../types';
 
 class Style {
+  private STYLE_LOADER = getResolvePath('style-loader');
   private CSS_LOADER = getResolvePath('css-loader');
   private SASS_LOADER = getResolvePath('sass-loader');
   private LESS_LOADER = getResolvePath('less-loader');
@@ -43,22 +44,11 @@ class Style {
       enableCssModule && this.sassCssModuleConfig,
       {
         test: /\.less$/,
-        exclude: [this.realAntdV4Path],
         use: [
           this.styleLoader,
           this.cssLoader,
           this.postCSSLoader,
           this.lessLoader,
-        ],
-      },
-      {
-        test: /\.less$/,
-        include: [this.realAntdV4Path],
-        use: [
-          this.styleLoader,
-          this.cssLoader,
-          this.postCSSLoader,
-          this.antdV4LessLoader,
         ],
       },
     ];
@@ -87,16 +77,9 @@ class Style {
     };
   }
 
-  //TODO: remove when upgrade to antd v4
-  get realAntdV4Path() {
-    const antdV4Path = join(this.opts.cwd, 'node_modules/antd-v4');
-    const ret = existsSync(antdV4Path) ? realpathSync(antdV4Path) : antdV4Path;
-    return ret;
-  }
-
   get styleLoader() {
     return {
-      loader: MiniCssExtractPluginLoader,
+      loader: this.opts.isProd ? MiniCssExtractPluginLoader : this.STYLE_LOADER,
     };
   }
 
@@ -125,17 +108,6 @@ class Style {
       options: {
         sourceMap: true,
         lessOptions,
-      },
-    };
-  }
-
-  get antdV4LessLoader() {
-    const { antdV4LessOptions = {} } = this.opts;
-    return {
-      loader: this.LESS_LOADER,
-      options: {
-        sourceMap: true,
-        lessOptions: antdV4LessOptions,
       },
     };
   }