+++ packages/auto-polyfills-webpack-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # auto-polyfills-webpack-plugin
 
+## 1.1.0
+
+### Minor Changes
+
+- fba9a04c: support new option
+  - auto-polyfills-webpack-plugin: support longTermCache option to decide whether or not to store cache polyfills.
+
+## 1.0.2
+
+### Patch Changes
+
+- 55aa11c9: fix bug
+  - auto polyfills webpack plugin can't read browserslist correctly
+
 ## 1.0.1
 
 ### Patch Changes